requirements.txt
# python =3.9.18
pandas
torch
transformers==4.38.2
gdown==5.1.0
peft
evaluate
sentence_transformers
matplotlib
google-generativeai
openai
rouge_score
sentencepiece


# cuda
#pip install torch==1.13.0+cu116 torchvision==0.14.0+cu116 torchaudio==0.13.0 --extra-index-url https://download.pytorch.org/whl/cu116


#allennlp
# pip install spacy==3.7.4
# python -m spacy download en_core_web_sm
# pip install --pre allennlp-models

#After downloading allennlp, you need to redownload cuda, transformers and openai.


